Remove Fullstory
Removed fullstory and transferred some common utilities over to the report web files file.
