+++ package.json
@@ -7,7 +7,6 @@
     "@actnowcoalition/assert": "^1.0.0",
     "@actnowcoalition/number-format": "^1.0.0",
     "@actnowcoalition/time-utils": "^1.0.0",
-    "@fullstory/browser": "^1.4.6",
     "@iconify/icons-twemoji": "^1.0.11",
     "@iconify/react": "^1.1.3",
     "@material-ui/core": "^4.11.0",
